Replace running total gauge metrics with counters
This adds 4 new counter metrics for read/write disk bytes and
receive/transmite network bytes. These counters duplicate the
functionality of existing gauge metrics, which are marked as deprecated.

Per prometheus docs, "a gauge is a metric that represents a single
numerical value that can arbitrarily go up and down" and a counter " is
a cumulative metric that represents a single monotonically increasing
counter".